+++ Packages/com.texelsaur.access/Runtime/Scripts/Validation/DataValidator.cs
@@ -1,0 +1,26 @@
+﻿
+using UdonSharp;
+using UnityEngine;
+using VRC.SDKBase;
+using VRC.Udon;
+
+namespace Texel
+{
+    public abstract class DataValidator : UdonSharpBehaviour
+    {
+        public virtual bool _PreValidate(string data)
+        {
+            return true;
+        }
+
+        public virtual string _Transform(string data)
+        {
+            return data;
+        }
+
+        public virtual bool _PostValidate(string data)
+        {
+            return true;
+        }
+    }
+}