+++ Packages/com.texelsaur.access/Runtime/Scripts/Validation/DataValidatorKey.cs
@@ -1,0 +1,13 @@
+﻿
+using UdonSharp;
+using UnityEngine;
+using VRC.SDKBase;
+using VRC.Udon;
+
+namespace Texel
+{
+    public abstract class DataValidatorKey : UdonSharpBehaviour
+    {
+        public abstract string _GetKey();
+    }
+}